image_scraper/requirements.txt
requests
tqdm
Pillow
torch
transformers
accelerate
bitsandbytes
